requirements.txt
# IntelliResume AI – Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
openai>=1.12.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pydantic>=2.5.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
